requirements.txt
distlib==0.3.9
filelock==3.18.0
greenlet==3.4.0
numpy==2.4.4
pandas==3.0.2
platformdirs==4.3.8
playwright==1.58.0
prompt_toolkit==3.0.52
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
virtualenv==20.31.2
wcwidth==0.6.0
